# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API del sistema hospitalario Galenos Pro. Arquitectura hexagonal, persiste en SQL Server 2022.",
        "title": "Galenos Pro Appointments API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/triaje": {
            "get": {
                "description": "Devuelve los triajes del período filtrado (SP ListarTriaje_Emergencia). El filtro de texto se aplica en el SP; si no se envía, se pasa vacío.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triaje"
                ],
                "summary": "Lista los triajes registrados",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha inicial (YYYY-MM-DD)",
                        "name": "fini",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha final (YYYY-MM-DD)",
                        "name": "ffin",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Texto de búsqueda",
                        "name": "filtro",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Servicio de derivación (-100 para todos)",
                        "name": "derivadoAServicio",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Estado del triaje",
                        "name": "idEstado",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al listar los triajes",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Persiste el triaje del paciente invocando el SP webTab_PacienteTriajeAgregar. Devuelve el @Resultado informado por el procedimiento.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triaje"
                ],
                "summary": "Registra un triaje de paciente",
                "parameters": [
                    {
                        "description": "Datos del triaje a registrar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpadapter.createTriajeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado informado por el SP",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Cuerpo inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al registrar el triaje",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/triaje/admision": {
            "post": {
                "description": "Crea la atención (admisión) de un paciente desde su triaje invocando el SP WebCrearAtencionDesdeTriaje. Devuelve el @Resultado informado por el procedimiento.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triaje"
                ],
                "summary": "Admisiona un paciente con triaje",
                "parameters": [
                    {
                        "description": "Datos de la admisión a partir del triaje",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpadapter.createAdmissionFromTriageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado informado por el SP",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Cuerpo inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al admisionar al paciente",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/triaje/consulta/atencion/{idAtencion}": {
            "get": {
                "description": "Devuelve el triaje vigente (UltimoTriaje = 1) de la atención indicada, o null si no existe.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triaje"
                ],
                "summary": "Obtiene el triaje de consulta externa de una atención",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id de la atención",
                        "name": "idAtencion",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al obtener el triaje",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpadapter.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpadapter.apiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/httpadapter.apiError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "httpadapter.createAdmissionFromTriageRequest": {
            "type": "object",
            "required": [
                "idPacienteTriaje",
                "idTriaje"
            ],
            "properties": {
                "direccionPaciente": {
                    "type": "string"
                },
                "idMedico": {
                    "type": "integer"
                },
                "idPacienteTriaje": {
                    "type": "integer"
                },
                "idTriaje": {
                    "type": "integer"
                },
                "nombreAcompanante": {
                    "type": "string"
                },
                "observacion": {
                    "type": "string"
                },
                "telefonoAcompanante": {
                    "type": "string"
                }
            }
        },
...
    }
}